+++ src/app/data/game-state.js
@@ -281,8 +281,9 @@ module.exports = {
    * @method newPlayer
    */
   newPlayer: function() {
+    this.numExtraLives = 0;
     this.score = 0;
-    this.fuel = 9000;
+    this.fuel = 8000;
     this.lives = 5;
   },
 
@@ -329,8 +330,8 @@ module.exports = {
       return;
     }
     if (this.bonuses.orbRecovered && !this.isGameOver) {
-      if (levelManager.levels.length - 1 === levelManager.levelIndex) {
-        //this.newGame();
+      if (levelManager.levels.length - 1 === levelManager.levelIndex &&
+        !options.gameModes.endlessMode.enabled) {
         this.levelsCompleted.dispatch();
       } else {
         this.levelStart();
@@ -389,6 +390,32 @@ module.exports = {
     planetBuster: false,
     orbRecovered: false
   },
+
+  setScore: function(value) {
+    this.score = value;
+    this.check1Up();
+  },
+
+  addScore: function(value) {
+    this.score += value;
+    this.check1Up();
+  },
+
+  /**
+   * @method check1Up
+   */
+  check1Up: function() {
+    if (this.score / (10000 * (this.numExtraLives + 1)) >= 1) {
+      console.log('1 UP!', this.score / (10000 * (this.numExtraLives + 1)));
+      this.lives += 1;
+      this.numExtraLives++;
+    }
+  },
+
+  /**
+   *
+   */
+  numExtraLives: 0,
 
   /**
    * @property score