+++ src/app/environment/Collisions.js
@@ -1,5 +1,9 @@
+var levelManager = require('../data/level-manager');
+
 /**
- * Collisions description
+ * Helper class to initialise Collision groups
+ * also starts the p2 physics system
+ * Implements reverse gravity in endless running game mode
  *
  * @class Collisions
  * @constructor
@@ -20,8 +24,7 @@ p.startSystem = function () {
   game.physics.startSystem(Phaser.Physics.P2JS);
   game.physics.p2.useElapsedTime = true;
   game.physics.p2.setImpactEvents(true);
-  game.physics.p2.gravity.y = 100;
-
+  game.physics.p2.gravity.y = levelManager.endlessData.flip? -100 : 100;
 };
 
 /**
@@ -39,7 +42,6 @@ p.init = function () {
   this.enemies = game.physics.p2.createCollisionGroup();
   this.fuels = game.physics.p2.createCollisionGroup();
   this.drones = game.physics.p2.createCollisionGroup();
-
   game.physics.p2.updateBoundsCollisionGroup();
 };