+++ src/app/ui/mission-swipe.js
@@ -55,10 +55,33 @@ module.exports = {
     this.swipe.height = 5;
     this.group.add(this.swipe);
     this.group.fixedToCamera = true;
+
+    /**
+     "title": "Mission %n",
+     "desc": "Recover the orb %s %s",
+     "color": "rgba(255, 0, 0, 0.7)"
+     *
+     * @type {{font: string, fill: string, align: string}}
+     */
+
+    console.log('levelIndex=', levelManager.levelIndex);
+    console.log('levels.length=', levelManager.levels.length);
+    console.log('levels.endlessModeIndex=', levelManager.endlessModeIndex);
+    console.log('levels.endlessCycle=', levelManager.endlessCycle);
+    var levelNo = levelManager.levelIndex + 1 +
+      levelManager.levels.length * levelManager.endlessModeIndex +
+      levelManager.endless.length * levelManager.endlessCycle;
+    console.log('levels.levelNo = ', levelNo);
+    var title = level.missionSwipe.title.replace('%n', levelNo.toString());
+    var repl1 = levelManager.endlessData.flip? "\nReverse Gravity" : "";
+    var repl2 = levelManager.endlessData.blink? "\nBlink Map" : "";
+    var desc;
+    desc = level.missionSwipe.desc.replace('%s1', repl1);
+    desc = desc.replace('%s2', repl2);
     style = { font: "24px thrust_regular", fill: "#ffffff", align: "left" };
-    this.title = game.add.text(x + 5, y + 5, level.missionSwipe.title, style, this.group);
+    this.title = game.add.text(x + 5, y + 5, title, style, this.group);
     style = { font: "12px thrust_regular", fill: "#ffffff", align: "left" };
-    this.desc = game.add.text(this.title.x, this.title.y + this.title.height + 2, level.missionSwipe.desc, style, this.group);
+    this.desc = game.add.text(this.title.x, this.title.y + this.title.height + 2, desc, style, this.group);
     this.hideSwipe();
     this.tl = new TimelineMax({delay: 0.2, onComplete: this.missionStartSwipeOut, callbackScope: this, onReverseComplete: this.missionReady});
     this.tl.add(TweenMax.to(this.swipe, 0.2, {alpha: 1, ease: Quad.easeOut} ));